Cast time to unsigned value for random seed

+++ main.c
@@ -40,7 +40,7 @@ main(int argc, char *argv[])
     (void)argc;
     (void)argv;
 
-    CHECK_RC(snk_create_field(15, 10, SNK_ARRAY_LEN(obstacles), obstacles, time(NULL),
+    CHECK_RC(snk_create_field(15, 10, SNK_ARRAY_LEN(obstacles), obstacles, (uint32_t)time(NULL),
             &field));
 
     CHECK_RC(snk_create(&field, &start_position, SNK_RIGHT, 5, &process));
